requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
python-multipart==0.0.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.4.0
